backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
elevenlabs==1.4.0
indic-transliteration==2.3.57
googletrans==4.0.0rc1